test(copilot): Add vitest runner and make normalizeVariantForProperty testable in isolation
Wires up vitest (+happy-dom) for @umbraco-ai/agent-copilot; the existing
tsconfig.json exclude for src/**/*.test.ts already covers the renamed test file.

Extracts normalizeVariantForProperty into its own dependency-free module
(variant-normalization.ts) so its unit tests don't have to load the full
CMS/backoffice module graph pulled in transitively by property-value-tool-base.ts.
property-value-tool-base.ts re-exports it for backward compatibility.
