+++ Umbraco.AI.Agent.Copilot/src/Umbraco.AI.Agent.Copilot/Client/package.json
@@ -11,7 +11,8 @@
         "watch": "tsc && vite build --watch",
         "build": "npm run build:api && npm run build:app",
         "build:app": "tsc -p tsconfig.json && vite build",
-        "build:api": "tsc -p tsconfig.api.json && (api-extractor run -c api-extractor.json --verbose || exit 0)"
+        "build:api": "tsc -p tsconfig.api.json && (api-extractor run -c api-extractor.json --verbose || exit 0)",
+        "test": "vitest run"
     },
     "dependencies": {
         "@umbraco-ai/agent": "*",
@@ -21,7 +22,9 @@
     },
     "devDependencies": {
         "@microsoft/api-extractor": "^7.55.1",
+        "happy-dom": "^20.10.6",
         "typescript": "^5.9.2",
-        "vite": "^8.0.16"
+        "vite": "^8.0.16",
+        "vitest": "^4.1.10"
     }
 }
