+++ Umbraco.AI.Agent.Copilot/src/Umbraco.AI.Agent.Copilot/Client/vitest.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+    test: {
+        environment: "happy-dom",
+        include: ["src/**/*.test.ts"],
+    },
+});
