+++ package.json
@@ -27,6 +27,7 @@
         "watch:agent": "npm run watch -w @umbraco-ai/agent",
         "watch:agent-ui": "npm run watch -w @umbraco-ai/agent-ui",
         "watch:copilot": "npm run watch -w @umbraco-ai/agent-copilot",
+        "test:copilot": "npm run test -w @umbraco-ai/agent-copilot",
         "generate-client": "concurrently --names \"core,prompt,agent\" -c \"blue,green,yellow\" \"npm run generate-client:core\" \"npm run generate-client:prompt\" \"npm run generate-client:agent\"",
         "generate-client:core": "npm run generate-client -w @umbraco-ai/core",
         "generate-client:prompt": "npm run generate-client -w @umbraco-ai/prompt",
